% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_geoid_crosswalk}
\alias{acs_geoid_crosswalk}
\title{ACS Processing/Merging Functions}
\usage{
acs_geoid_crosswalk(
  df,
  acs_tract = "acs_tract",
  acs_block = "acs_block",
  state_fips = "state_fips",
  county_fips = "county_fips",
  census_tract = "census_tract",
  census_block = "census_block"
)
}
\arguments{
\item{df}{Dataframe for crosswalk.}

\item{acs_tract}{Variable name for ACS tract geoids.}

\item{acs_block}{Variable name for ACS block geoids.}

\item{state_fips}{Variable name for state FIPS codes.}

\item{county_fips}{Variable name for county FIPS codes.}

\item{census_tract}{Variable name for Census Geocoder tract geoids.}

\item{census_block}{Variable name for Census Geocoder block geoids.}
}
\description{
These functions are not for export.
}
