\name{attachHTML}
\alias{attachHTML}
\title{Attach Common, Non-Masking HTML Functions}
\usage{
  attachHTML()
}
\description{
  Calling this function will attach HTML tagging functions
  that do not mask any functions available in the 'base' R
  packages.
}
\seealso{
  \code{\link{detachHTML}}
}

