\name{awk.set}
\alias{awk.set}
\title{Set awk}
\usage{
awk.set(awk)
}
\arguments{
  \item{awk}{String denoting the appropriate call for your
  flavour of \code{awk}.}
}
\description{
Use this function to set the string by which \code{awk} is
called; e.g. if you're using GNU awk (gawk), mawk, and so
on.
}

