% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLSD.R
\name{calcLSD}
\alias{calcLSD}
\alias{plotLSD}
\title{Approximate pairwise least significant interval from an asreml object.}
\usage{
calcLSD(asrpred, weight = NULL)

plotLSD(asrpred, weight = NULL, order = c("no", "increasing", "decreasing"),
  ...)
}
\arguments{
\item{asrpred}{Result of a \code{predict.asreml(..., sed=TRUE)$prediction} call.}

\item{weight}{A weight matrix with the same dimension as the \code{sed} matrix. It feeds into the \code{lm} call which weighs the \code{sed} matrix.}

\item{order}{One of \code{"no"}(default), \code{"increasing"}, or \code{"decreasing"}. Whether the resulting plot should be ordered by the predicted mean.}
}
\description{
It approximate a pariwise LSD from the sed matrix from an asreml predict call. The goal is to ease plotting.
}
\examples{
\donttest{
data(oats)
oats.asr <-
  asreml(yield ~ Variety * Nitrogen,
         random = ~ Blocks / Wplots,
         data = oats)
oats.pred <-
  predict(oats.asr, classify = "Variety", sed = TRUE)$prediction
plotLSD(oats.pred)
plotLSD(oats.pred, order = "inc")
plotLSD(oats.pred, order = "dec")
calcLSD(oats.pred)
}
}

