\name{cat.cb}
\alias{cat.cb}
\title{Write Data to the Clipboard}
\usage{
  cat.cb(dat, ...)
}
\arguments{
  \item{dat}{data to be written to the clipboard.}

  \item{...}{optional arguments passed to \code{cat}.}
}
\description{
  This function writes data to the clipboard, using
  \code{\link{cat}}.
}
\seealso{
  \code{\link{write.cb}}, \code{\link{cat}}
}

