% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_format.R
\name{census_format}
\alias{census_format}
\title{Format Various Geocoding Outputs into Census Format}
\usage{
census_format(
  df,
  input_type = "n/a",
  longitude = "longitude",
  latitude = "latitude",
  street = "street",
  city = "szSitusCity",
  state = "sSitusState",
  zip = "sSitusZip",
  cutoff = 0.87,
  accuracy = NULL,
  addr_all = NULL,
  sep = ", ",
  file_date = NULL,
  vars_keep = NULL
)
}
\arguments{
\item{df}{Input dataframe.}

\item{input_type}{Type of output. You can specify `geocodio` or `dstk`.}

\item{longitude}{Variable name for longitude. Defaults to "longitude."}

\item{latitude}{Variable name for latitude. Defaults to "latitude"}

\item{street}{Variable for street address.}

\item{city}{Variable for city.}

\item{state}{Variable for state abbreviation.}

\item{zip}{Variable for zip code.}

\item{cutoff}{If specified with `accuracy`, will filter for geocoded
outputs with accuracy level above the cutoff. Defaults to 0.87.}

\item{accuracy}{Variable for accuracy.}

\item{addr_all}{Variable for one-line address, or street-city-state-zip
pasted. It will be created if not specified as existing variable.}

\item{sep}{Separator to use in creating `addr_all`.}

\item{file_date}{Geocoded date. Defaults to `Sys.time()`.}

\item{vars_keep}{Variables to keep besides the census format.}
}
\description{
This function takes a dataframe of geocoding outputs and wrangles it
into a Census Geocoder output format for comparability. The main variable
names are set as "address_input", "match", "match_detail", "address_output",
"latlon", "tiger", "side", "state_fips", "county_fips", "census_tract",
and "census_block".
}
