% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vars.R
\name{clean_vars}
\alias{clean_vars}
\title{Variable Cleaning}
\usage{
clean_vars(df = NULL, varnames = NULL, varnames_date = NULL,
  date_order = "mdy", varnames_num = NULL, firstname = NULL,
  prefix = NULL, gender_original = NULL)
}
\arguments{
\item{df}{Dataframe to be cleaned}

\item{varnames}{All variables to be cleaned}

\item{varnames_date}{Date variables}

\item{date_order}{Order of the date variable, if string format}

\item{varnames_num}{Numeric variables}

\item{firstname}{Variable containing first names}

\item{prefix}{Variable containing self-reported personal prefixes}

\item{gender_original}{Variable containing original gender entry}
}
\description{
This function takes a dataframe and performs a variety of cleaning.
Characters are stripped of unnecessary whitespaces, non-ASCII, and
lowercased. Numbers are stripped of non-numeric and turned to numeric.
Date variables are parsed. Finally, gender is inferred if first name
variable is provided, overwritten by self-reported prefixes and original
gender variable, if it exists. Function title_gender_infer is a variation
of this function.
}
