% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.allFit}
\alias{coef.allFit}
\title{Exract coefficients from ensemble object.}
\usage{
\method{coef}{allFit}(object, ...)
}
\arguments{
\item{object}{Output from \code{fitAll}.}
}
\value{
A data.frame with coefficients.
}
\description{
It exract coefficients from ensemble object.The code borrows heavily from package \code{meifly}.
}
\examples{
\donttest{
allreg=fitAll(conc~uptake*Treatment,CO2)
coef(allreg)
}
}

