% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.ensemble}
\alias{coef.ensemble}
\title{Exract coefficients from ensemble object.}
\usage{
\method{coef}{ensemble}(object, ...)
}
\arguments{
\item{object}{Ensemble object. Output from \code{\link{fitall}}}
}
\value{
A data.frame with coefficients.
}
\description{
It exract coefficients from ensemble object.The code borrows heavily from package \code{meifly}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ensemble}: 
}}
\examples{
\donttest{
allreg=fitall(conc~uptake*Treatment,CO2)
coef(allreg)
}
}

