% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_mismatch.R
\name{col_mismatch}
\alias{col_mismatch}
\title{Return Rows With Different Values Between Columns}
\usage{
col_mismatch(df, cols = NULL, ref = NULL)
}
\arguments{
\item{df}{Dataframe to be cleaned.}

\item{cols}{cols
Defaults to NULL.}

\item{ref}{ref}
}
\description{
This function takes a dataframe and if two columns are specified,
returns rows with different values between these columns, including
instances where one is an NA value.
}
\details{
If no columns are specified, it checks for instances of `name_repair` given
by the three-dots (https://principles.tidyverse.org/names-attribute.html;
see Section 3.5.4 "Ugly, with a purpose").
}
