% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas_files.R
\name{cp_dat}
\alias{cp_dat}
\title{Copy all of the files in a directory into another directory}
\usage{
cp_dat(dir_in, dir_out, ...)
}
\arguments{
\item{dir_in}{The directory with files to copy}

\item{dir_out}{The directory to copy into. If this doesn't exist it will be created.}

\item{...}{Additional arugments passed to fs::dir_ls()}
}
\description{
Copies files from one directory into another, only copying files that have been updated.
}
