% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv.R
\name{createSCV}
\alias{createSCV}
\title{Create Stratified (Cross-)Validation Set}
\usage{
createSCV(form, data, k = 5)
}
\arguments{
\item{form}{A \code{formula}. The LHS of \code{form} (if exists) and all non-factors in the RHS will be removed before forming strata.}

\item{data}{Data.}

\item{k}{number of folds to form.}

\item{valid_prob}{Percentage of data to appear in validation set. Only used in \code{createSV}.}
}
\value{
A numeric vector with entries indicating the fold it belongs to.
}
\description{
It creates stratified (cross-)validation folds from data according to formula, so that all levels of all factors in \code{formula} are as balancely scattered as possible. 
This avoids problem with simple un-stratified cross validation where a factor level presents only in the validation set but not in the training set.
}
\examples{
\donttest{
createSCV(~Type+Treatment,CO2)
createSV(~Type+Treatment,CO2)

set.seed(123)
f1=createSCV(~Type+Treatment,CO2)
set.seed(123)
f2=createSCV(conc~Type+Treatment+uptake,CO2)
identical(f1,f2) ## TRUE
}
}

