% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvAll.R, R/vall.R
\name{cvAll}
\alias{cvAll}
\alias{cvAll.default}
\alias{cvAll.lm}
\alias{cvAll.lme}
\alias{cvAll.merMod}
\alias{validateAll}
\title{Cross Validate all nested models}
\usage{
cvAll(x, ...)

\method{cvAll}{default}(x, data, k, lossfn = .calcMSE, ind, finalfit = TRUE)

\method{cvAll}{lm}(x, data, k = 5, ind = NULL, ...)

\method{cvAll}{lme}(x, data, k = 5, ind = NULL, ...)

\method{cvAll}{merMod}(x, data, k = 5, ind = NULL, ...)

validateAll(x, data, method, valid_prob, ...)
}
\arguments{
\item{x}{Either an \code{lm} object or a two side \code{\link[stats]{formula}}.}

\item{data}{Data.}

\item{k}{Postive integer. k-fold cross validation is carried out. Defaults to 5.}

\item{lossfn}{A function with 2 arguments. The first takes an array of predicted value, the second takes a vector of observed value. 
Then it calculates the loss. 
It defaults to a function which calculates the mean sqaured error (MSE).}

\item{finalfit}{if \code{TRUE} models are refitted to the entire dataset at the end.}

\item{valid_prob}{Positive number between 0 and 1. 
It is the proportion of data used in the validation set.}
}
\value{
An maxtrix of predictions on the validation set, with the following attributes:

\code{loss} The loss calculated by \code{lossfn}.

\code{data} The data used.

\code{models} The fitted models or just formulae of models, depending on \code{finalfit}.
}
\description{
It fits all subset regression, allowing interaction and categorical covariates, and cross validates them.
The code borrows heavily from package \code{meifly}.
}
\examples{
## Using cvAll
\donttest{
allreg=cvAll(conc~uptake*Treatment,CO2)
allreg
attr(allreg,'models') 

calcmae=function(pred,obs)mean(abs(pred-obs),na.rm=TRUE)
allreg=cvAll(conc~uptake*Treatment,CO2,lossfn=calcmae)

## Using cvAll with class lm
lm0=lm(conc~uptake*Treatment,CO2)
allreg2=cvAll(lm0,CO2)

}
}

