% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{Return a Deduplicated a Dataframe}
\usage{
dedup(df, vars = NULL)
}
\arguments{
\item{df}{Dataframe to be deduplicated.}

\item{vars}{Variables to use in deduplication.}
}
\description{
This function takes a dataframe and returns only the deduplicated rows,
based on all columns.
}
