\name{df2list}
\alias{df2list}
\title{Convert data.frame to list}
\usage{
df2list(df, inplace = FALSE)
}
\arguments{
  \item{df}{A data.frame.}

  \item{inplace}{Boolean. If \code{TRUE}, we convert the
  list in place, so that the \code{list} itself is
  transformed into a \code{data.frame}, sans copying.}
}
\description{
This function converts a \code{data.frame} to a list.
}

