% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dict}
\alias{dict}
\title{Python-style Dictionaries in R}
\usage{
dict(..., `_size` = 29L)
}
\arguments{
\item{...}{Named arguments used in constructing the dictionary.}

\item{`_size`}{The number of 'buckets' used. This is the maximum of
\code{29L} and the number of named arguments passed to \code{...}.}
}
\description{
Dictionaries, i.e., hashed key-value pairs, are implemented in \R using
environments as a backend.
}
\details{
Dictionaries are just hashed \R environments with \code{emptyenv()} as a
parent.
}
\section{Warning}{


Dictionaries have \bold{reference semantics}, so modifying a dictionary
within a function will modify the dictionary passed in, not a copy! Use the
\code{copy} function to duplicate a \code{dict}.
}
\examples{
## Reference semantics -- be careful!
x <- dict()
y <- x
x[["a"]] <- 100
print(y[["a"]])

## Use copy to be explicit
y <- copy(x)
x[["b"]] <- 200
try(y[["b"]], silent = TRUE)

## Named lookup can be much faster in a dictionary
x <- as.list(1:1E5)
names(x) <- paste0("Element_", 1:1E5)
dict <- as.dictionary(x)
if (require(microbenchmark)) {
  microbenchmark(
    x[["Element_1"]],
    dict[["Element_1"]],
    x[["Element_1000"]],
    dict[["Element_1000"]],
    x[["Element_100000"]],
    dict[["Element_100000"]]
  )
}
}

