% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_dendrogram.R
\name{dir_dendrogram}
\alias{dir_dendrogram}
\title{Directory Tree Dendrogram}
\usage{
dir_dendrogram(path)
}
\arguments{
\item{path}{List of directories to be converted to a dendrogram}
}
\description{
This function takes a vector of directory structures and turns it into a
dendrogram. The code is a variation of Tyler Rinker's StackOverflow post
here: https://stackoverflow.com/questions/36094183/how-to-build-a-dendrogram-from-a-directory-tree
Note that you can print everything by such as following:
`print(dir_dendrogram(list.dirs()), limit = 300)``
}
