% Generated by roxygen2 (4.0.0): do not edit by hand
\name{error-wrappers}
\alias{error-wrappers}
\alias{message_if}
\alias{messagef}
\alias{stop_if}
\alias{stopf}
\alias{warn_if}
\alias{warnf}
\title{Stop / Warning / Message Wrappers}
\usage{
stopf(fmt, ...)

warnf(fmt, ...)

messagef(fmt, ...)

stop_if(expr, fmt, ...)

warn_if(expr, fmt, ...)

message_if(expr, fmt, ...)
}
\arguments{
\item{expr}{An expression to be evaluated and checked for 'truthiness'.}

\item{fmt}{A character vector of format strings. Passed to \code{\link{gettextf}}.}

\item{...}{Optional arguments passed to \code{\link{gettextf}}.}
}
\description{
These are simple wrappers to functions \code{stop}, \code{warning}, and
\code{message}.
}

