% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_addr}
\alias{ex_addr}
\title{Examples of U.S. Residential Addresses}
\format{A dataframe with 1,000 rows and 5 variables:
\describe{
  \item{street}{Street address}
  \item{city}{City}
  \item{state}{State}
  \item{zip}{5-digit zip code}
  \item{address_input}{First four variables pasted, separated by comma}
}}
\usage{
ex_addr
}
\description{
A dataset containing 1,000 example residential addresses of the U.S.,
sampled from the Federal Election Commission receipts, Schedule A,
individuals, within the 2018 Election Cycle. 48 states and DC addresses
present, with MS and ND obs missing, presumably due to small sample.
All entries are stripped of non-ASCII, punctuations, extra spaces or tabs,
and all are lowercased.
}
\keyword{datasets}
