\name{factor_to_char}
\alias{factor_to_char}
\title{Converts Factors to Characters in an Object}
\usage{
  factor_to_char(x)
}
\arguments{
  \item{x}{an object.}
}
\description{
  Converts factors to characters in an object. Leaves
  non-factor elements untouched. Objects that can be
  subscripted with \code{[[} are supported.
}

