\name{fast_factor}
\alias{fast_factor}
\title{Fast Factor Generation}
\usage{
  fast_factor(x)
}
\arguments{
  \item{x}{An object of atomic type \code{integer},
  \code{numeric}, \code{character} or \code{logical}.}
}
\description{
  This function generates factors quickly using faster
  sorting and matching algorithms available in Rcpp.
}

