% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAll.R
\name{fitAll}
\alias{fitAll}
\alias{fitAll.lm}
\alias{fitAll.lme}
\alias{fitAll.merMod}
\title{All subset regression}
\usage{
fitAll(x, ...)

\method{fitAll}{lm}(x, data, description = "fitting", progress = TRUE, ...)

\method{fitAll}{lme}(x, data, description = "fitting", progress = TRUE, ...)

\method{fitAll}{merMod}(x, data, description = "fitting", progress = TRUE,
  ...)
}
\arguments{
\item{x}{A model \code{object} of one of the classes supported.}

\item{data}{Data.}

\item{description}{Text to print for progress bar.}

\item{progress}{\code{TRUE/FALSE}. If TRUE progress bar is shown.}
}
\value{
A list with n components. n is the number of models fitted.
}
\description{
It fits all subset regression, allowing interaction and categorical covariates. The code borrows heavily from package \code{meifly}.
}
\examples{
## Using fitAll with formula class
\donttest{
lm0=lm(conc~uptake*Treatment,CO2)
allreg=fitAll(lm0,CO2)
}
}

