% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_mutate_df.R
\name{gender_mutate_df}
\alias{gender_mutate_df}
\title{Inferring Gender from First Names}
\usage{
gender_mutate_df(df, input_name, years = c(1932, 2012), method = "ssa",
  countries = "United States")
}
\arguments{
\item{df}{Dataframe containing a column of first names}

\item{input_name}{The name of the column containing first names in given df}

\item{years}{Which years to use for inferring gender: defaults to 1932-2012.}

\item{method}{Which method to use for inferring gender: defaults to "ssa".}

\item{countries}{Which country to use as a baseline: defaults to the U.S.}
}
\description{
This function uses ropensci/gender to convert first names in a dataframe
and obtain additional columns from the ropensci/gender outcome.
This function complements ropensci/gender to be directly used with dataframes
}
\examples{
df <- data.frame(NameFirst = c("Madison", "Nimrodel", "Michael"), stringsAsFactors = FALSE)
gender_mutate_df(df, "NameFirst", years = c(1920, 2010))
}
