% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_curl.R
\name{geocoder_curl_input}
\alias{geocoder_curl_input}
\title{Create Census Geocoder Input Files and Curl Submission File}
\usage{
geocoder_curl_input(
  df,
  input_path = ".",
  input_prefix = NULL,
  output_path = ".",
  output_prefix = NULL,
  curl_path = ".",
  curl_file = NULL,
  curl_append = FALSE,
  benchmark = "Public_AR_Census2010",
  vintage = "Census2010_Census2010"
)
}
\arguments{
\item{df}{Input dataframe.}

\item{input_path}{File path for input text files.}

\item{input_prefix}{File prefix for input text files.}

\item{output_path}{File path for output text files.}

\item{output_prefix}{File prefix for output text files.}

\item{curl_path}{File path for curl command file.}

\item{curl_file}{File name for curl command file, without the extension.}

\item{curl_append}{Whether to append an existing curl file.}

\item{benchmark}{The benchmark selection for Census Geocoder.}

\item{vintage}{The vintage selection for Census Geocoder.}
}
\description{
This function takes a dataframe and creates a series of 10,000 line input
text files to be uploaded to the Census Geocoder for batch geocoding.
It also creates a text file in which curl commands are written.
}
