% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder_curl.R
\name{geocoder_curl_output}
\alias{geocoder_curl_output}
\title{Create Census Geocoder Input Files and Curl Submission File}
\usage{
geocoder_curl_output(
  df,
  output_path = ".",
  output_prefix = NULL,
  benchmark = "Public_AR_Census2010",
  vintage = "Census2010_Census2010"
)
}
\arguments{
\item{df}{Input dataframe.}

\item{output_path}{File path for output text files.}

\item{output_prefix}{File prefix for output text files.}

\item{benchmark}{The benchmark selection for Census Geocoder.}

\item{vintage}{The vintage selection for Census Geocoder.}
}
\description{
This function takes a dataframe and creates a series of 10,000 line input
text files to be uploaded to the Census Geocoder for batch geocoding.
It also creates a text file in which curl commands are written.
}
