% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Rt.R
\name{get_Rt}
\alias{get_Rt}
\title{Simulate Rt}
\usage{
get_Rt(t)
}
\arguments{
\item{t}{Time \code{$\beta$} - transmission rate per unit of time,}
}
\value{
A vector of \code{$R_t$}
}
\description{
The function \code{stoch_seir} implements a stochastic SEIR model that includes three
parameters: transmission rate per unit time ($\beta$), average incubation
period (1/$\sigma$), and recovery rate ($\gamma$).
}
