% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAllForm.R
\name{getAllForm}
\alias{getAllComb}
\alias{getAllForm}
\title{Get All nested model formulae/combination matrix.}
\usage{
getAllForm(x)

getAllComb(x)
}
\arguments{
\item{x}{A formula.}
}
\value{
\code{getAllForm} returns a list of formulae.

\code{getAllComb} returns of variable combination matrix.
}
\description{
getAllForm find all nested models' formulae.
getAll FixedComb represent all nested models in terms of a combination matrix.
}
\details{
The nested models are formed under one rule: Interaction terms are present only if main effects are.
}

