% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getObjects}
\alias{getObjects}
\title{Get all Objects in Environment}
\usage{
getObjects(env)
}
\arguments{
\item{env}{an environment.}
}
\value{
a list of the objects contained within that environment.
}
\description{
Get all objects within an environment. Useful for inspecting the objects
available in a particular environment.
}
\examples{
myenv <- new.env()
assign( "foo", "bar", env=myenv )
assign( "baz", "spam", env=myenv )
getObjects( myenv )
}

