% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie_run.R
\name{gillespie_run}
\alias{gillespie_run}
\title{Simulate the model implemented in the Gillespie's direct method}
\usage{
gillespie_run(func, tend, nrun, y0, params, report_dt = 1)
}
\arguments{
\item{tend}{Simulation time}

\item{nrun}{The number of simulation runs}

\item{y0}{Initial states}

\item{params}{Parameters for the model}

\item{fun}{Model}
}
\value{
Simulation results
}
\description{
Simulate the model implemented in the Gillespie's direct method
}
