% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie_seir.R
\name{gillespie_seir}
\alias{gillespie_seir}
\title{An Gillespie algorithm (direct method) implementation of a stochastic SEIR model}
\usage{
gillespie_seir(y, params)
}
\arguments{
\item{params}{Parameters \code{$beta$} - transmission rate per unit of time,
\code{sigma} rate of transitioning from exposed to infectious state,
\code{gamma} rate of transitioning from infectious to recovered state}
}
\value{
A vector of changes in S, E, I, and R
}
\description{
The function \code{gillespie_seir} implements a stochastic SEIR model based on #' a Gillespie algorithm
parameters: transmission rate per unit time ($beta$), average incubation
period (1/$sigma$), and recovery rate ($gamma$).
}
