% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie_sepair.R
\name{gillespie_sepair}
\alias{gillespie_sepair}
\title{An Gillespie algorithm (direct method) implementation of a stochastic
SEPAIR model
The function \code{gillespie_sepair} implements a stochastic SEPAIR model
based on #' a Gillespie direct method
parameters: transmission rate per unit time ($beta$), mean incubation
period (1/$delta$), mean latent period (1/$epsilon$)
and recovery rate ($gamma$).}
\usage{
gillespie_sepair(y, params)
}
\arguments{
\item{params}{Parameters \code{$beta$} - transmission rate per unit of time,
\code{delta} rate of transitioning from exposed to onset of symptoms (E->P->I)
\code{gamma} rate of transitioning from infectious to recovered state}
}
\value{
A vector of changes in S, E, I, and R
}
\description{
An Gillespie algorithm (direct method) implementation of a stochastic
SEPAIR model
The function \code{gillespie_sepair} implements a stochastic SEPAIR model
based on #' a Gillespie direct method
parameters: transmission rate per unit time ($beta$), mean incubation
period (1/$delta$), mean latent period (1/$epsilon$)
and recovery rate ($gamma$).
}
