% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{grid_arrange_shared_axes}
\alias{grid_arrange_shared_axes}
\title{Common-Legend Arrangement of Plots}
\usage{
grid_arrange_shared_axes(
  ...,
  list = NULL,
  ncol = NULL,
  nrow = 1,
  widths = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  fontfamily = "CM Roman",
  pdf_default = TRUE
)
}
\arguments{
\item{...}{A series of plots to share common axes.}

\item{list}{A list of plots to share common axes.
Defaults to NULL. If list is specified, ellipsis argument is ignored.}

\item{ncol}{Number of columns of the arranged plots.
Defaults to NULL, in which case it is the length of the list.}

\item{nrow}{Number of rows of the arranged plots}

\item{widths}{Width to control the plot size.}

\item{title}{Plot-specific titles. Defaults to NULL.}

\item{xlab}{Common x-axis title to plots. Defaults to first plot's.}

\item{ylab}{Common y-axis title to plots. Defaults to first plot's.}

\item{fontfamily}{Font family to be used in common axes labels.
Defaults to CM Roman.}

\item{pdf_default}{Whether to apply pdf_default, which can reset `theme`.}
}
\description{
This function creates a common legend for the given pots to share
This was taken from the following link:
https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
