% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_race.R
\name{impute_race}
\alias{impute_race}
\title{Impute Race to Categorical Variables After wru::predict_race}
\usage{
impute_race(
  df,
  census,
  state = "sSitusState",
  surname = "szNameLast",
  county = "county_fips",
  tract = "census_tract",
  block = "census_block",
  id = "lVoterUniqueID",
  keep_probs = FALSE,
  ...
)
}
\arguments{
\item{df}{Dataframe prepped for wru::predict_race.}

\item{census}{Census object.}

\item{state}{State variable.}

\item{surname}{Surname variable.}

\item{county}{County variable.}

\item{tract}{Tract variable.}

\item{block}{Block variable.}

\item{id}{ID variable.}

\item{keep_probs}{Whether to keep the underlying probabilities generating
the classification.}

\item{...}{Other arguments to be passed to wru::predict_race.}
}
\description{
If you want to pass age, sex, and party, pass it exactly as in wru
so that the ellipsis captures it. Detailed documentation will follow soon.
}
