% Generated by roxygen2 (4.0.0): do not edit by hand
\name{install_rforge}
\alias{install_rforge}
\title{Attempts to install a package directly from R-Forge.}
\usage{
install_rforge(project, ...)
}
\arguments{
\item{project}{project name}

\item{...}{Other arguments passed on to \code{\link{install.packages}}.}
}
\description{
Attempts to install a package directly from R-Forge.
}
\examples{
\dontrun{
install_rforge("data.table")
}
}

