\name{kCharToFactor}
\alias{kCharToFactor}
\title{Converts Characters to Factors in Data Frame}
\usage{
  kCharToFactor(x, ...)
}
\arguments{
  \item{x}{a \code{data.frame} or \code{list} object}

  \item{...}{passed to \code{\link{factor}} function}
}
\description{
  Converts characters to factors in a \code{data.frame}.
  Leaves non-character columns untouched.
}

