% Generated by roxygen2 (4.0.0): do not edit by hand
\name{kImg}
\alias{kImg}
\title{HTML - Source an Image}
\usage{
kImg(x, width = 480, height = 480)
}
\arguments{
\item{x}{path to an image you want to source}

\item{width}{width (in pixels) of the image}

\item{height}{height (in pixels) of the image}
}
\description{
Convenience function for \code{cat}-ing out HTML markup for an image as
\code{<img src="..." />}.
}
\seealso{
\code{\link{hImg}}
}

