% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{label_df}
\alias{label_df}
\title{Extract variable labels}
\usage{
label_df(.data)
}
\arguments{
\item{.data}{A \code{data.frame}}
}
\value{
A \code{data.frame} with the attributes \code{names} and
   \code{variable.labels} as columns.
}
\description{
Extract variable labels
}

