% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list2df}
\alias{list2df}
\title{Convert list to data.frame}
\usage{
list2df(list, inplace = FALSE)
}
\arguments{
  \item{list}{A list.}

  \item{inplace}{Boolean. If \code{TRUE}, we convert the
  list in place, so that the \code{list} itself is
  transformed into a \code{data.frame}, sans copying.}
}
\description{
This function converts a list to a data frame, assuming that each
element of the list is of equal length.
}

