\name{list2mat}
\alias{list2mat}
\title{Convert a list to a matrix}
\usage{
list2mat(df)

list2mat(df)
}
\arguments{
  \item{list}{A list.}

  \item{df}{a \code{data.frame}.}
}
\description{
This function converts a \code{list} to a \code{matrix},
assuming that each element of the list is of equal length.

This function converts a \code{data.frame} to a
\code{matrix}.
}

