% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Load RData into a Pre-Specified Assignment}
\usage{
loadRData(file_name)
}
\arguments{
\item{file_name}{The file name of the .RData object.}
}
\description{
This function is from [StackOverflow post 5577221](https://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file) here,
Ricardo's answer.
It loads an Rda file and enables it to be assigned to a specific object,
instead of using the object name that it was stored with.
This is an incredibly useful function that deserves a place in base R,
but until then I will use it here.
}
