% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_dummy}
\alias{make_dummy}
\title{Make Dummy Variables from a Factor}
\usage{
make_dummy(x)
}
\arguments{
  \item{x}{an object coercable to \code{factor}.}
}
\description{
This functions converts a single factor into dummy variables, with one
dummy variable for each level of that factor. Names are constructed as
\code{<varName>_<level>}.
}
\examples{
x <- factor( rep( c("a", "b", "c", "d"), each=25 ) )
make_dummy(x)
}

