% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode Function}
\usage{
Mode(x, na.rm = T)
}
\arguments{
\item{x}{The input vector to compute the mode}

\item{na.rm}{Whether to remove NA values. Defaults to TRUE.}
}
\description{
This is a function that outputs a mode of a sgiven vector
This is named Mode instead of mode to prevent conflict with base::mode
This was taken from Stackoverflow:
https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
\examples{
x <- c(1, 2, 2, 2, 3, 3, NA)
Mode(x)

}
\keyword{mode}
