% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nan_replace.R
\name{nan_replace}
\alias{nan_replace}
\title{Check NaN values in a dataframe and replace them}
\usage{
nan_replace(df, replace = TRUE, value = 0)
}
\arguments{
\item{df}{Dataframe to be checked for NaN values}

\item{replace}{Whether to replace NaN values.}

\item{value}{Value to replace NaN with.}
}
\description{
This function takes a dataframe and checks for NaN values.
It also replaces the NaN values if requested, a with user-specified value.
This function was partially taken from the following StackOverflow link:
https://stackoverflow.com/questions/18142117/how-to-replace-nan-value-with-zero-in-a-huge-data-frame/18143097
}
\examples{
df <- data.frame(a = c(1, 2), b = c(NaN, 3))
nan_replace(df)
}
