% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_seir.R
\name{ode_seir}
\alias{ode_seir}
\title{An implementation of a stochastic SEIR model}
\usage{
ode_seir(t, y, params)
}
\arguments{
\item{params}{Parameters \code{$\beta$} - transmission rate per unit of time,
\code{$\sigma$} rate of transitioning from exposed to infectious state,
\code{$\gamma$} rate of transitioning from infectious to recovered state}
}
\value{
A list of changes in S, E, I, and R
}
\description{
This function implements a stochastic SEIR model that includes three
parameters: transmission rate per unit time \code{$\beta$}, average incubation
period \code{1/$\sigma$}, and recovery rate \code{$\gamma$}.
}
