% Generated by roxygen2 (4.0.0): do not edit by hand
\name{par.reset}
\alias{par.reset}
\title{Restore the 'par' settings}
\usage{
par.reset()
}
\description{
If you have been mucking around with parameter settings, this function
will revert to the parameter settings that were available when this
package was loaded.
}

