% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_na.R
\name{paste_na}
\alias{paste_na}
\title{Paste with NA suppression}
\usage{
paste_na(..., sep = " ")
}
\arguments{
\item{...}{List of R objects, to be converted to character vectors.}

\item{sep}{A character string to separate the terms.}
}
\description{
This enables `paste` to work even when some of the entries are NA values,
and the desired output is an empty string when NAs are pasted.
}
\details{
This is originally from the following StackOverflow post:
https://stackoverflow.com/questions/13673894/suppress-nas-in-paste#15673180
This is simply a cosmetically changed version of 42-'s answer.
}
