% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_default.R
\name{pdf_default}
\alias{pdf_default}
\title{Graphics Option: PDF Articles, Default Option}
\usage{
pdf_default(p, CMRoman = NULL)
}
\arguments{
\item{p}{The input ggplot2 object}

\item{CMRoman}{Font for the graphics. Default is CM Roman.}
}
\description{
This function implements my *personal* favorite setup for article pdfs.
This does not currenty contain details on colors.
Note that the CM Roman font will not show immediately when called in the
Rstudio plot pane, but will be correctly input in the resulting pdf file.
}
\examples{
library(ggplot2)
p <- ggplot(diamonds, aes(depth, colour = cut)) + geom_density()
pdf_default(p)
}
\keyword{plot}
