% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_nogrid}
\alias{plot_nogrid}
\title{No Grid/Axis Plots}
\usage{
plot_nogrid(p)
}
\arguments{
\item{p}{The plot that we will strip its grids/axes of.}
}
\description{
This is a function that strips the plot of all its grid and axes elements.
}
