% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_nolegend}
\alias{plot_nolegend}
\title{No Legend Plots}
\usage{
plot_nolegend(p)
}
\arguments{
\item{p}{The plot that we will strip its legend of.}
}
\description{
This is a function that strips the plot of its legend.
}
\examples{
library(ggplot2)
p <- ggplot(diamonds, aes(depth, colour = cut)) +
  geom_density()
plot_nolegend(p)
}
