% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_notitle}
\alias{plot_notitle}
\title{No Title Plots}
\usage{
plot_notitle(p)
}
\arguments{
\item{p}{The plot that we will strip its titles of.}
}
\description{
This is a function that strips the plot of all its titles,
including axis titles and legend titles.
}
\examples{
library(ggplot2)
p <- ggplot(diamonds, aes(depth, colour = cut)) + geom_density() +
  ggtitle("Diamonds")
plot_notitle(p)
}
