% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numeric.R
\name{plot_numerics}
\alias{plot_numerics}
\title{Autmated histograms}
\usage{
plot_numerics(.data, .labels = NULL, missingness = T)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{.labels}{A second \code{data.frame}, containing the variables with
corresponding labels to set titles}
}
\value{
Plots as side effect.
}
\description{
This functions returns histograms for all numeric variables in a data set.
}

