% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_default.R
\name{png_default}
\alias{png_default}
\title{Graphics Option: PNG Presentations, Default Option}
\usage{
png_default(p, OfficinaSans = NULL)
}
\arguments{
\item{p}{The input ggplot2 object}

\item{OfficinaSans}{Font for the graphics. Default is OfficinaSansITBoo.}
}
\description{
This function implements my *personal* favorite setup for beamer pngs.
This does not currenty contain details on colors.
}
\examples{
library(ggplot2)
p <- ggplot(diamonds, aes(depth, colour = cut)) + geom_density()
png_default(p)
}
\keyword{plot}
