% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_condprob.R
\name{pretty_condprob}
\alias{pretty_condprob}
\title{Append Row Number as a Variable to a Dataframe}
\usage{
pretty_condprob(
  df,
  A_var,
  A_val,
  B_var,
  B_val,
  output = "string",
  digits = 1,
  percent = TRUE
)
}
\arguments{
\item{df}{Input dataframe.}

\item{A_var}{The A variable in "the conditional probability of A given B."}

\item{A_val}{Value of A variable in interest.}

\item{B_var}{The B variable in "the conditional probability of A given B."}

\item{B_val}{Value of B variable in interest.}

\item{output}{Whether to print a numeric number or a statement.
If the argument is "string", the function prints a message containing
the conditional probability.
Defaults to "string."}

\item{digits}{The digit of the final output. Defaults to 1.}

\item{percent}{Whether to display the conditional probability in a
percentage. Defaults to TRUE.#'}
}
\description{
This function takes a dataframe and adds a variable of row numbers.
This is the simplest function that does not take into account any group
structures.
}
