% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_skeleton.R
\name{proj_skeleton}
\alias{proj_skeleton}
\title{Create R Project Skeleton}
\usage{
proj_skeleton(path = ".")
}
\arguments{
\item{path}{The directory in which to create the skeletal directories.}
}
\description{
Creates R project skeleton. This is not an R package skeleton.
}
