% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Rescale Numeric Vector to [0, 1] range}
\usage{
range01(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector to be rescaled.}

\item{na.rm}{A logical value indicating whether to remove NA values before computation. Default is FALSE.}
}
\value{
A rescaled numeric vector.
}
\description{
This function takes a numeric interval
and rescales it to be between 0 and 1.
}
