\name{rcpp_apply_generator}
\alias{rcpp_apply_generator}
\title{Rcpp apply Generator}
\usage{
  rcpp_apply_generator(fun, name = NULL, file = NULL)
}
\arguments{
  \item{fun}{A character string defining the C++ function.
  It must be in terms of a variable \code{x}, and it must
  return a \code{double}. \code{x} is a reference to the
  current row/column being iterated over.}

  \item{name}{An internal name for the function.}

  \item{file}{A location to output the file. Defaults to a
  temporary file as generated by \code{\link{tempfile}}().}
}
\description{
  Use this function as a generator for your own
  \code{apply} functions; that is, functions you would like
  to apply over rows or columns of a matrix.
}

