% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas_files.R
\name{read_sas_files}
\alias{read_sas_files}
\title{Read in all of the SAS files in a directory}
\usage{
read_sas_files(directory)
}
\arguments{
\item{directory}{The directory containing \code{.sas7bdat} files.}
}
\description{
Read in all of the SAS files in a directory
}
