\name{remove_chars}
\alias{remove_chars}
\title{Remove Alphabetic Characters from a Character Vector}
\usage{
remove_chars(x, remove_spaces = TRUE)
}
\arguments{
  \item{x}{A character vector, or vector coercable to
  character.}

  \item{remove_spaces}{boolean; if \code{TRUE} we remove
  all white-space as well.}
}
\description{
Removes all alphabetic characters from a character vector.
}

