\name{remove_digits}
\alias{remove_digits}
\title{Remove Digits from a Character Vector}
\usage{
remove_digits(x, remove_spaces = TRUE)
}
\arguments{
  \item{x}{A character vector, or vector coercable to
  character.}

  \item{remove_spaces}{boolean; if \code{TRUE} we remove
  all white-space as well.}
}
\description{
Removes all digits from a character vector.
}

