% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_seeds.R
\name{rep_seeds}
\alias{rep_seeds}
\title{Create Fully Reproducible Seeds}
\usage{
rep_seeds(folds = 10, n = 10000, seed = 123)
}
\arguments{
\item{folds}{Number of folds for cross-validation. Defaults to 10/}

\item{n}{Number of seeds to set per fold. Defaults to 10,000.}

\item{seed}{Initial seed. Defaults to 123.}
}
\description{
Create fully reproducible seeds for cross-validation in some applications
of machine learning.
}
