% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_seq.R
\name{row_seq}
\alias{row_seq}
\title{Append Row Number as a Variable to a Dataframe}
\usage{
row_seq(df, row = "row")
}
\arguments{
\item{df}{Input dataframe.}

\item{row}{Name for the row variable.}
}
\description{
This function takes a dataframe and adds a variable of row numbers.
This is the simplest function that does not take into account any group
structures.
}
