% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowid_matrix_to_df.R
\name{rowid_matrix_to_df}
\alias{rowid_matrix_to_df}
\title{Matrix to Dataframe with Row ID Preservation}
\usage{
rowid_matrix_to_df(x, colname = "rownames")
}
\arguments{
\item{x}{The object(s) with row IDs to be converted to dataframes}

\item{colname}{The column name that will contain the row names}
}
\description{
Similar to tibble::rowid_to_column but also converting matrix to dataframe.
It can also take lists of matrix/dataframes.
}
\examples{
mdat <- matrix(c(1, 2, "A", "B"), nrow = 2, ncol = 2, byrow = TRUE,
               dimnames = list(c("row1", "row2"), c("col1", "col2")))
rowid_matrix_to_df(mdat, colname = "row")
}
