% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_format.R
\name{scale_format}
\alias{scale_format}
\title{Change Decimal Places in Axis Labels}
\usage{
scale_format(x, format = "\%.1f")
}
\arguments{
\item{x}{Character to format}

\item{format}{How to reformat the labels}
}
\value{
A reformatted character
}
\description{
This function change the number of decimal places on axis labels.
This was taken from the following link:
https://stackoverflow.com/questions/38722202/how-do-i-change-the-number-of-decimal-places-on-axis-labels-in-ggplot2
}
