% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleFUN.R
\name{scaleFUN}
\alias{scaleFUN}
\title{Change Decimal Places in Axis Labels}
\usage{
scaleFUN(x, format = "\%.1f")
}
\arguments{
\item{x}{Object to format}

\item{format}{How to reformat the labels}
}
\description{
This function change the number of decimal places on axis labels.
This was taken from the following link:
https://stackoverflow.com/questions/38722202/how-do-i-change-the-number-of-decimal-places-on-axis-labels-in-ggplot2
}
