% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_switch.R
\name{selenium_switch}
\alias{selenium_switch}
\title{Switch Window Handles in RSelenium}
\usage{
selenium_switch(remDr, windowID)
}
\arguments{
\item{remDr}{`remoteDriver` object created from RSelenium `rsDriver`.}

\item{windowID}{ID of the window handle to switch to.}
}
\description{
This function is from [RSelenium GitHub Issue #143](https://github.com/ropensci/RSelenium/issues/143):
John Harrison's answer.
This helps switch between window handles when using RSelenium.
Call such as `remDr$getWindowHandles()[[1]]` and so on.
}
