% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setnamed}
\alias{setnamed}
\title{Set the NAMED attribute}
\usage{
setnamed(x, i)
}
\arguments{
  \item{x}{An \R object.}

  \item{i}{An integer equal to 1 or 2.}
}
\description{
This function sets the \code{NAMED} attribute of an \R object.
For development and debugging purposes only.
}

