% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setrownames}
\alias{setcolnames}
\alias{setrownames}
\title{Set the row, column names of a matrix in place}
\usage{
setrownames(x, value)

setcolnames(x, value)
}
\arguments{
  \item{x}{Either a \code{data.frame}, or an array.}

  \item{value}{Either \code{NULL} or a vector coercible to
  character.}
}
\description{
Similar to the \code{setattr} or \code{setnames} functions of
\code{data.table}, but makes it slightly easier to set these attributes
in-place for matrices.
}

