% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Simple Capitalization}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{This is a character or a vector of characters}
}
\value{
A character or a vector of characters with first char capitalized.
}
\description{
This function takes a string and outputs
a string with only the first character capitalized.
This was taken from the following link:
https://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string
}
\examples{
simple_cap("attila marcel")
}
