\name{split_}
\alias{split_}
\title{Fast Split}
\usage{
split_(x, f)
}
\arguments{
  \item{x}{A vector containing values to be divided into
  groups.}

  \item{f}{A grouping variable defining the grouping.}
}
\description{
A faster version of \code{\link{split}} -- this divides the
data in the vector \code{x} into the groups defined by
\code{f}.
}

