% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splom2.R
\name{splom2}
\alias{splom2}
\title{Modified lattice::splom.}
\usage{
splom2(x, grid = T, pscale = 0, type = c("r", "g", "p"), col.line = 2,
  noi = NULL, noj = noi, ...)
}
\arguments{
\item{x, grid, pscale, type, col.line, ...}{Same as that in \code{\link[lattice]{splom}}.}

\item{noi}{Panels in row \code{i} are not plotted.}

\item{noj}{Panels in column \code{j} are not plotted.}
}
\description{
Upper panel computes correlation, and lower panel adds regression line.
}
\examples{
\donttest{
splom2(~iris[1:4])
splom2(~iris[1:4],noj=4,noi=4)
}
}

