% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_trans.R
\name{squish_trans}
\alias{squish_trans}
\title{Squishing Y-Axis}
\usage{
squish_trans(from, to, factor)
}
\arguments{
\item{from}{Lower bound of the interval to be squished}

\item{to}{Upper bound of the interval to be squished}

\item{factor}{factor to be used in squishing-interval}
}
\value{
A trans object that can be used to ggplot
}
\description{
This function takes an interval and outputs a trans object
that "squishes" the y-axis where it may be uninformative.
This was taken from the following link:
https://stackoverflow.com/questions/35511951/r-ggplot2-collapse-or-remove-segment-of-y-axis-from-scatter-plot
}
