% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_df}
\alias{state_df}
\title{State Names and Abbreviations Dataframe}
\format{
A dataframe with 51 rows and 2 variables:
\describe{
  \item{abb}{State Abbreviations}
  \item{names}{State Names}
}
}
\usage{
state_df
}
\description{
A dataset that puts state.abb and state.names together into a single
dataframe, along with DC.
}
\keyword{datasets}
