% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoch_seir.R
\name{stoch_seir}
\alias{stoch_seir}
\title{An implementation of a stochastic SEIR model}
\usage{
stoch_seir(y, params, delta)
}
\arguments{
\item{params}{Parameters \code{$\beta$} - transmission rate per unit of time,
\code{$\sigma$} rate of transitioning from exposed to infectious state,
\code{$\gamma$} rate of transitioning from infectious to recovered state}
}
\value{
A vector of changes in S, E, I, and R
}
\description{
The function \code{stoch_seir} implements a stochastic SEIR model that includes three
parameters: transmission rate per unit time ($\beta$), average incubation
period (1/$\sigma$), and recovery rate ($\gamma$).
}
