% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoch_solve.R
\name{stoch_solve}
\alias{stoch_solve}
\title{A routine to extract outputs from a stochastic SEIR model}
\usage{
stoch_solve(func, y, times, params, delta)
}
\arguments{
\item{func}{Function such as a stochstic SEIR model}

\item{times}{A sequence of times for which you want to see the output}

\item{params}{Parameter values for the stochastic model}
}
\value{
A list of changes in S, E, I, and R
}
\description{
This function implements a stochastic SEIR model that includes three
parameters: transmission rate per unit time ($\beta$), average incubation
period (1/$\sigma$), and recovery rate ($\gamma$).
}
