% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisFile.R
\name{thisFile}
\alias{thisFile}
\title{Locating R Script File}
\usage{
thisFile()
}
\description{
This function locates the running script's location.
The code is from steamer25's StackOverflow post
here: https://stackoverflow.com/questions/1815606/rscript-determine-path-of-the-executing-script
This enables both sourcing and Rscript-ing.
}
