% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_infer.R
\name{title_gender_infer}
\alias{title_gender_infer}
\title{Inferring Gender from Personal Titles}
\usage{
title_gender_infer(df, input_name, output_name = "title_gender")
}
\arguments{
\item{df}{Dataframe containing a column of titles}

\item{input_name}{The name of the column containing titles in the given df}

\item{output_name}{The name of the column that will contain inferred gender}
}
\description{
This function takes a dataframe with a column of personal titles,
such as Ms. Miss. Mrs. Mr.
and outputs another column with respectively inferred gender.
}
\examples{
title_gender_infer(data.frame(title = c("Mr.", "Mrs")), "title")
}
\keyword{gender}
