% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tree}
\alias{tree}
\title{Print a Tree Representation of an Object of Nested Lists}
\usage{
tree(x)
}
\arguments{
\item{x}{A (named) list.}
}
\description{
This function returns output similar to that of the command line
tool \code{tree}, except rather than directory/file structure,
we simply print the names of lists.
}
\seealso{
http://stackoverflow.com/questions/18122548/display-names-of-column-of-recursive-list-as-tree
}

