% Generated by roxygen2 (4.0.0): do not edit by hand
\name{u}
\alias{u}
\title{Unique elements in a vector}
\usage{
u(...)
}
\arguments{
\item{...}{passed to \code{\link{unique}}.}
}
\description{
Returns the unique elements in a vector. A wrapper to
\code{\link{unique}()}.
Primarily intended for interactive, not programmatic, use.
}

