% Generated by roxygen2 (4.0.0): do not edit by hand
\name{us}
\alias{us}
\title{unlist( strsplit( ... ) )}
\usage{
us(x, split = "", ...)
}
\arguments{
\item{x}{vector of items, as passed to \code{\link{strsplit}}}

\item{split}{the delimiter to split on}

\item{...}{optional arguments passed to strsplit}
}
\description{
This is a thin wrapper to \code{ unlist( strsplit( ... ) ) }.
Primarily intended for interactive, not programmatic, use.
}
\examples{
x <- "apple_banana_cherry"
us(x, "_")
}
\seealso{
\code{\link{unlist}}, \code{\link{strsplit}}
}

