% Generated by roxygen2 (4.0.0): do not edit by hand
\name{value_matching}
\alias{\%drop\%}
\alias{\%keep\%}
\alias{\%kin\%}
\alias{\%knin\%}
\alias{\%nin\%}
\alias{value_matching}
\title{Value Matching}
\usage{
x \%nin\% y

x \%kin\% y

x \%knin\% y

x \%drop\% y

x \%keep\% y
}
\arguments{
  \item{x}{Vector or \code{NULL}: the values to be
  matched.}

  \item{y}{Vector or \code{NULL}: the values to be matched
  against.}
}
\description{
These are a couple of mostly self-explanatory wrappers around \code{\%in\%}.
}
\details{
\code{\%nin\%} returns a logical vector indicating if there is
no match for its left operand. It is the inverse of \code{x \%in\% y}.

\code{\%kin\%} returns the actual values of \code{x} for which
\code{x \%in\% y}.

\code{\%knin\%} returns the actual values of \code{x} for which
\code{x \%nin\% y}.
}

