% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_clean.R
\name{zip_clean}
\alias{zip_clean}
\title{Clean Zip Codes}
\usage{
zip_clean(df, zip_var = NULL, keep_last_four = FALSE, last_four_label = NULL)
}
\arguments{
\item{df}{Input dataframe or vector.}

\item{zip_var}{Variable that contains zip codes, if df is dataframe.}

\item{keep_last_four}{Whether to keep the last four digits.}

\item{last_four_label}{Variable to store last four digits of zip, if any.}
}
\description{
This function takes a dataframe or vector containing zip codes and convert
them to the standard five-digit zip codes. Depending on the argument, it
saves the subdigits.
}
