\name{char_to_factor}
\alias{char_to_factor}
\title{Converts Characters to Factors in an Object}
\usage{
  char_to_factor(X, inplace = FALSE, ...)
}
\arguments{
  \item{X}{an object.}

  \item{inplace}{boolean; if \code{TRUE} the object is
  modified in place. Be careful when using this option.}

  \item{...}{Ignored.}
}
\description{
  Converts characters to factors in an object. Leaves
  non-factor elements untouched.
}

