\name{char_to_factor}
\alias{char_to_factor}
\title{Converts Characters to Factors in an Object}
\usage{
  char_to_factor(X, ...)
}
\arguments{
  \item{X}{an object.}

  \item{...}{optional arguments passed to \code{factor}.}
}
\description{
  Converts characters to factors in an object. Leaves
  non-factor elements untouched.
}

