% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vars.R
\name{clean_vars}
\alias{clean_vars}
\title{Variable Cleaning}
\usage{
clean_vars(df, varnames = NULL, varnames_date = NULL,
  date_order = "mdy", varnames_num = NULL, firstname = NULL,
  gender = TRUE, prefix = NULL, prefix_male = "mr",
  gender_original = NULL, gender_male = "^m$|^male$")
}
\arguments{
\item{df}{Dataframe to be cleaned.}

\item{varnames}{All variables to be cleaned.
Defaults to NULL.}

\item{varnames_date}{Date variables.
Defaults to NULL.}

\item{date_order}{Order of the date variable, if string format.
Defaults to "mdy".}

\item{varnames_num}{Numeric variables.
Defaults to NULL.}

\item{firstname}{Variable containing first names.
Defaults to NULL.}

\item{gender}{Whether to create an inferred gender variable.
Defaults to TRUE.}

\item{prefix}{Variable containing self-reported personal prefixes.
Defaults to NULL.}

\item{prefix_male}{Value that indicates male in the prefix variable.
Defaults to "mr" (the input will be lowercased before comparison.)}

\item{gender_original}{Variable containing original gender entry.
Defaults to NULL.}

\item{gender_male}{Regex expression that indicates male
in the original gender entry. Defaults to "^m$|^male$"
(the input will be lowercased before comparison.)}
}
\description{
This function takes a dataframe and performs a variety of cleaning.
Characters are stripped of unnecessary whitespaces, non-ASCII, and
lowercased. Numbers are stripped of non-numeric and turned to numeric.
Date variables are parsed. Finally, gender is inferred if first name
variable is provided, overwritten by self-reported prefixes and original
gender variable, if it exists. Function title_gender_infer is a variation
of this function.
}
