% Generated by roxygen2 (4.0.0): do not edit by hand
\name{counts}
\alias{counts}
\title{Generate Counts of Values in a Vector}
\usage{
counts(x)
}
\arguments{
  \item{x}{A numeric, integer, character or logical vector,
  or a (potentially nested) list of such vectors. If
  \code{x} is a list, we recursively apply \code{counts}
  throughout elements in the list.}
}
\description{
This function uses Rcpp sugar to implement a fast \code{table}, for
unique counts of a single vector. This implementation seeks to
produce identical output to \code{table(x, useNA="ifany")}.
}
\details{
The order of \code{NA}, \code{NaN} in the output may differ -- even
\R is inconsistent with the order that \code{NA} and \code{NaN} elements
are inserted.
}
\examples{
x <- round( rnorm(1E2), 1 )
counts(x)
}

