% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dict}
\alias{dict}
\title{Python-style Dictionaries in R}
\usage{
dict(..., `_size` = 29L)
}
\arguments{
\item{...}{Named arguments used in constructing the dictionary.}

\item{`_size`}{The number of 'buckets' used. This is the maximum of
\code{29L} and the number of named arguments passed to \code{...}.}
}
\description{
Dictionaries, i.e., hashed key-value pairs, are implemented in \R using
environments as a backend.
}
\details{
Dictionaries are just hashed \R environments with \code{emptyenv()} as a
parent.
}
\examples{
## Named lookup can be much faster in a dictionary
x <- as.list(1:1E5)
names(x) <- paste0("Element_", 1:1E5)
dict <- as.dictionary(x)
if (require(microbenchmark)) {
  microbenchmark(
    x[["Element_1"]],
    dict[["Element_1"]],
    x[["Element_100000"]],
    dict[["Element_100000"]]
  )
}
}

