% Generated by roxygen2 (4.0.0): do not edit by hand
\name{enumerate}
\alias{enumerate}
\title{Enumerate over a Vector}
\usage{
enumerate(X, FUN, ...)
}
\arguments{
\item{X}{A vector.}

\item{FUN}{A function, taking two arguments.}

\item{...}{Optional arguments to \code{FUN}.}
}
\description{
This function extends \code{lapply} to operate on functions taking two
arguments -- the first refers to each element within a vector, while
the second refers to the current index.
}
\examples{
data <- setNames( list(1, 2, 3), c('a', 'b', 'c') )
v <- replicate(10, rnorm(1E3), simplify=FALSE)
identical( lapply(v, sum), enumerate(v, sum) )
f <- function(x, i) i
enumerate(v, f)
enumerate(v, function(x, i) i)
}

