% Generated by roxygen2 (4.0.0): do not edit by hand
\name{factor_to_char}
\alias{factor2char}
\alias{factor_to_char}
\title{Converts Factors to Characters in an Object}
\usage{
factor_to_char(X, inplace = FALSE)

factor2char(X, inplace = FALSE)
}
\arguments{
\item{X}{an object.}

\item{inplace}{Boolean; if \code{TRUE} we modify the object in place.
Useful if you're modifying a list and don't want to force a copy, but
be wary of other symbols pointing at the same data.}
}
\description{
Converts factors to characters in an object. Leaves non-character
elements untouched.
}
\details{
We iterate through all elements in the object (e.g. if
it is a list) and convert anything that is a factor into a character.
}

