% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\docType{data}
\name{fips}
\alias{fips}
\title{State-level Geographic Indicators}
\format{A data frame with 51 rows and 5 variables:
\describe{
  \item{stname}{State names, in lowercase}
  \item{stabb}{State abbreviations}
  \item{stfips}{FIPS code}
  \item{south_all}{Southern states}
  \item{south_conf}{Southern states (11 confederate)}
}}
\usage{
fips
}
\description{
A dataset containing the geographic indicators for U.S. states,
including the District of Columbia.
}
\keyword{datasets}
