% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_all.R
\name{fitAll}
\alias{fitAll}
\alias{fitAll.default}
\alias{fitAll.lm}
\title{All subset regression}
\usage{
fitAll(x, data, method, ...)

\method{fitAll}{default}(x, data, method = "lm", description = "fitting",
  progress = TRUE, ...)

\method{fitAll}{lm}(x, data, method = "lm", ...)
}
\arguments{
\item{x}{Either an \code{lm} object or a two side \code{\link[stats]{formula}}.}

\item{data}{Data.}

\item{method}{Currently must be 'lm'}
}
\value{
A list with n components. n is the number of models fitted.
The list also has an attribute:

\code{data} The data used in the fitting.
}
\description{
It fits all subset regression, allowing interaction and categorical covariates. The code borrows heavily from package \code{meifly}.
}
\examples{
## Using fitAll with formula class
\donttest{
allreg=fitAll(conc~uptake*Treatment,CO2)
allreg[[1]]
attr(allreg,'data')

## Using fitAll with lm class
lm0=lm(conc~uptake*Treatment,CO2)
allreg2=fitAll(lm0,CO2)
}
}

