% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Common-Legend Arrangement of Plots}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"), legend = NULL, legend_plot = NULL)
}
\arguments{
\item{...}{A series of plots to share a common legend}

\item{ncol}{Number of columns of the arranged plots}

\item{nrow}{Number of rows of the arranged plots}

\item{position}{Position of the legend}

\item{legend}{If legend is supplied from outside. Defaults to NULL.}

\item{legend_plot}{If plot to pull the legend from is supplied from outside.
Defaults to NULL.}
}
\description{
This function creates a common legend for the given pots to share
This was taken from the following link:
https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
