\name{kFactorToChar}
\alias{kFactorToChar}
\title{Converts Factors to Characters in Data Frame}
\usage{
  kFactorToChar(x)
}
\arguments{
  \item{x}{a \code{data.frame} or \code{list} object.}
}
\description{
  Converts factors to characters in a \code{data.frame}.
  Leaves non-factor columns untouched.
}

