% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lg}
\alias{lg}
\title{length( grep( ... ) )}
\usage{
lg(pattern, x, perl = TRUE, ...)
}
\arguments{
  \item{pattern}{regex pattern passed to \code{grep}.}

  \item{x}{a vector on which we attempt to match
  \code{pattern} on.}

  \item{perl}{boolean. use perl-compatible regular
  expressions?}

  \item{...}{additional arguments passed to
  \code{\link{grep}}.}
}
\description{
This is a wrapper to a \code{length( grep( ... ) )}. See examples for usage.
Primarily intended for interactive, not programmatic, use.
}
\examples{
x <- c("apple", "banana", "cherry")
if( lg( "^ap", x ) > 0 ) {
  print( "regular expression '^ap' found in 'x'" )
  }
}
\seealso{
\code{\link{re_exists}}
}

