% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list2mat}
\alias{list2mat}
\title{Convert a list to a matrix}
\usage{
list2mat(list)
}
\arguments{
  \item{list}{A list.}
}
\description{
This function converts a \code{list} to a \code{matrix}, assuming that each
element of the list is of equal length.
}

