% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_varimp.R
\name{pdf_varimp}
\alias{pdf_varimp}
\title{Return a Deduplicated a Dataframe}
\usage{
pdf_varimp(
  x,
  filename,
  labels = NULL,
  font = NULL,
  n_max = 10,
  size = 12,
  width = 8,
  height = 4
)
}
\arguments{
\item{x}{Model output from caret.}

\item{filename}{File name to export the PDF.}

\item{labels}{If separate labels are provided for features.
Defaults to NULL.}

\item{font}{Which font to use. If not specified, defaults to
Kievit Offc Pro.}

\item{n_max}{Number of top variables to show. Defaults to 10.}

\item{size}{Size of the strip.text.x element from `ggplot2::theme`.
Defaults to 12.}

\item{width}{Width of the PDF export. Defaults to 8.}

\item{height}{Height of the PDF export. Defaults to 4.}
}
\description{
This function takes a dataframe and returns only the deduplicated rows,
based on all columns.
}
