% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prepare_package}
\alias{prepare_package}
\title{Prepare Package}
\usage{
prepare_package(build = TRUE, check = TRUE, install = FALSE,
  copy.tarball = TRUE)
}
\arguments{
\item{build}{Build the package with \code{R CMD build}?}

\item{check}{Check the package with \code{R CMD check}?}

\item{install}{Install the package with \code{R CMD INSTALL}? Only
done if \code{build} is \code{TRUE} as well.}

\item{copy.tarball}{If \code{build && copy.tarball}, we copy the tarball
generated by \code{R CMD build} to the \code{"dist"} folder.}
}
\description{
This function prepares the package such that all the C / C++ source files
are concatenated into one file (for each source). This decreases
compilation time, and produces a tarball that can be used for submission
to CRAN.
}

