% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Create Percentage Tables}
\usage{
prop(df, vars, digits = 1)
}
\arguments{
\item{df}{Input dataframe.}

\item{vars}{Variables to be put into \code{table}.}

\item{digits}{Digits under the decimal point to show.
Defaults to 1.}
}
\description{
This function takes a dataframe and creates a table in percentages.
This is an enhanced version of \code{prop.table}.
}
