% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_changes.R
\name{seq_changes}
\alias{seq_changes}
\title{Show Graphically How an Attribute Changes Over Time}
\usage{
seq_changes(x, names = TRUE, sub_str = NULL, len = 2, gap = 15)
}
\arguments{
\item{x}{The sequence to be analyzed.}

\item{names}{Whether the sequence is named. Defaults to TRUE.}

\item{sub_str}{Whether to use a substring of the names of the sequence.
Defaults to NULL. If entered, this should be two integers in a vector.}

\item{len}{The length of separation in showing the changes.
Defaults to 2.}

\item{gap}{The length of separation from names to values.
Defaults to 15.}
}
\description{
This function takes a single row of a dataframe, a vector, or a list,
checks for changes in consecutive values over the sequence, and show at
which points the values change.
}
\examples{
seq_changes(
  c("A", "A", "B", "A", "A", "C", "C", "C", "C", "D"),
  names = FALSE, gap = 1
)

}
