% Generated by roxygen2 (4.0.0): do not edit by hand
\name{silent}
\alias{silent}
\title{Hide All Potential Output when Evaluating an Expression}
\usage{
silent(expr)
}
\arguments{
\item{expr}{An \R expression.}
}
\description{
This function hides any output a function might try to write
to \code{stdout}, \code{stderr}, to the \R console as a message, warning,
and so on.
}

